refactor: Move types to 'utils' and adapt to all components
