--- web/types/User.ts
@@ -1,14 +1,0 @@
-/**
- * User type from the back-end.
- */
-export type User = {
-  userID: string;
-  username: string;
-  email: string;
-  phoneNumber: string;
-  fullName: string;
-  isActive: string;
-  updatedAt: string;
-  createdAt: string;
-  role: 'admin' | 'user';
-};
